% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{asd.binary.file.path}
\alias{asd.binary.file.path}
\title{Return the full path to the ASD binary example file}
\usage{
asd.binary.file.path()
}
\value{
full path to the ASD binary example file
}
\description{
Return the full path to the ASD binary example file
}
