% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{asd.sco.binary.file.path}
\alias{asd.sco.binary.file.path}
\title{Return the full path to the ASD SCO binary example file}
\usage{
asd.sco.binary.file.path()
}
\value{
The full path to the ASD SCO binary example file
}
\description{
Return the full path to the ASD SCO binary example file
}
