% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_file_paths.R
\name{bruker.opus.binary.file.path}
\alias{bruker.opus.binary.file.path}
\title{Return the full path to the Bruker Opus Binary test file.}
\usage{
bruker.opus.binary.file.path()
}
\value{
full path to the Bruker Opus Binary test file
}
\description{
Return the full path to the Bruker Opus Binary test file.
}
