% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{bruker.opus.binary.test.abs.file.path}
\alias{bruker.opus.binary.test.abs.file.path}
\title{Return the full path to the Bruker Opus Binary absorption test file}
\usage{
bruker.opus.binary.test.abs.file.path()
}
\value{
The full path to the Bruker Opus Binary absorption test file
}
\description{
Return the full path to the Bruker Opus Binary absorption test file
}
