% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{csiro.scans.file.path}
\alias{csiro.scans.file.path}
\title{Return the full path to the CSIRO SCANS example file}
\usage{
csiro.scans.file.path()
}
\value{
The full path to the Hone Lab Red example file
}
\description{
Return the full path to the CSIRO SCANS example file
}
