% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{csv.file.path}
\alias{csv.file.path}
\title{Return the full path to the CSV example file}
\usage{
csv.file.path()
}
\value{
The full path to the CSV example file
}
\description{
Return the full path to the CSV example file
}
