% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{csv.test.file.path}
\alias{csv.test.file.path}
\title{Return the full path to the CSV test file}
\usage{
csv.test.file.path()
}
\value{
The full path to the CSV test file
}
\description{
Return the full path to the CSV test file
}
