% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraSelectionFunctions.R
\name{getSelectionFunction}
\alias{getSelectionFunction}
\title{Return a spectra selection function.}
\usage{
getSelectionFunction(selectionName)
}
\arguments{
\item{spectraSelectionName}{Name of spectra selection function to return}
}
\value{
A spectra selection function
}
\description{
Return a spectra selection function.
}
\examples{
getSelectionFunction('spectraSelectionName')
}
