% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraTransformations.R
\name{getTransformationFunction}
\alias{getTransformationFunction}
\title{Return a transformation function.}
\usage{
getTransformationFunction(transformationName)
}
\arguments{
\item{transformationName}{Transformation name of function to return}
}
\value{
A spectra transformation function
}
\description{
Return a transformation function.
}
\examples{
getTransformationFunction('transformationName')
}
