% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraTransformations.R
\name{getTransformationInfo}
\alias{getTransformationInfo}
\title{Return information about the available spectral transformations.}
\usage{
getTransformationInfo(type = NULL)
}
\arguments{
\item{type}{Transformation types to return}
}
\value{
A dataframe of spectral transformation information
}
\description{
Return information about the available spectral transformations.
}
