% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{hone.lab.red.reduced.file.path}
\alias{hone.lab.red.reduced.file.path}
\title{Return the full path to the Hone Lab Red Reduced example file}
\usage{
hone.lab.red.reduced.file.path()
}
\value{
The full path to the Hone Lab Red Reduced example file
}
\description{
Return the full path to the Hone Lab Red Reduced example file
}
