% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{nicolet.spa.file.path}
\alias{nicolet.spa.file.path}
\title{Return the full path to the Nicolet spa example file}
\usage{
nicolet.spa.file.path()
}
\value{
The full path to the Nicolet spa example file
}
\description{
Return the full path to the Nicolet spa example file
}
