% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_file_for_reading.R
\name{parse.hone.lab.red}
\alias{parse.hone.lab.red}
\title{Parse a raw Hone Lab Red spectra file containing multiple spectra}
\usage{
parse.hone.lab.red(path, out.directory)
}
\arguments{
\item{path}{Full path to the raw Hone Lab Red spectra file}

\item{out.directory}{Full path to the output directory}
}
\value{
A list of file names each containing 1 sample and 1 background record
}
\description{
Parses a raw Hone Lab Red spectra file containing multiple spectra into individual files containing a single spectra for reading in read.soilspec() function
}
