% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseRawFileForReading.R
\name{parseSCANS}
\alias{parseSCANS}
\title{Parse a raw SCANS spectra file containing multiple spectra}
\usage{
parseSCANS(path, outDirectory)
}
\arguments{
\item{path}{Full path to the raw SCANS spectra file}

\item{outDirectory}{Full path to the output directory}
}
\value{
A list of file names each containing 1 sample and 1 background record
}
\description{
Parses a raw SCANS spectra file containing multiple spectra into individual files containing a single spectra for reading in read.soilspec() function
}
