% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{perkin.elmer.sp.peir.file.path}
\alias{perkin.elmer.sp.peir.file.path}
\title{Return the full path to the Perkin Elmer PE IR example file}
\usage{
perkin.elmer.sp.peir.file.path()
}
\value{
The full path to the Perkin Elmer PE IR example file
}
\description{
Return the full path to the Perkin Elmer PE IR example file
}
