% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{perkin.elmer.sp.pepe.file.path}
\alias{perkin.elmer.sp.pepe.file.path}
\title{Return the full path to the Perkin Elmer PEPE example file}
\usage{
perkin.elmer.sp.pepe.file.path()
}
\value{
The full path to the Perkin Elmer PEPE example file
}
\description{
Return the full path to the Perkin Elmer PEPE example file
}
