% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilspec.R
\name{read.bruker.opus.binary}
\alias{read.bruker.opus.binary}
\title{Read a Bruker Opus Binary soil spectroscopy file.
This function should be used when the file does not have the expected ".0" suffix.
A precondition for correct functioning is that the file is of the expected type.}
\usage{
read.bruker.opus.binary(path)
}
\arguments{
\item{path}{Full path to the file}
}
\value{
A result list containing a file read status, a data.frame of
wavenumber-intensity pairs, a list of any available metadata,
instrument mode, units, whether wavenumbers are in descending order,
instrument origin, spectrum type;
status will be non-zero if file does not exist or cannot be read (1),
is of an unknown format (2) or some other error occurred (4).
}
\description{
Read a Bruker Opus Binary soil spectroscopy file.
This function should be used when the file does not have the expected ".0" suffix.
A precondition for correct functioning is that the file is of the expected type.
}
