% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilspec.R
\name{read.csv}
\alias{read.csv}
\title{Read a CSV spectroscopy file.
This function should be used when the file does not have the expected ".csv" suffix.
A precondition for correct functioning is that the file is of the expected type.}
\usage{
read.csv(
  path,
  is.absorbance = F,
  is.reflectance = F,
  is.transmittance = F,
  source.col.names = c("wavenumber", "intensity")
)
}
\arguments{
\item{path}{Full path to the file}

\item{is.absorbance}{Does the sample data correspond to absorbance?}

\item{is.reflectance}{Does the sample data correspond to reflectance?}

\item{is.transmittance}{Does the sample data correspond to transmittance?}

\item{source.col.names}{Vector of CSV column names to be read corresponding
to wavenumber and intensity data frame columns.}
}
\value{
A result list containing a file read status, a data.frame of
wavenumber-intensity pairs, a list of any available metadata,
instrument mode, units, whether wavenumbers are in descending order,
instrument origin, spectrum type;
status will be non-zero if file does not exist or cannot be read (1),
is of an unknown format (2) or some other error occurred (4).
}
\description{
Read a CSV spectroscopy file.
This function should be used when the file does not have the expected ".csv" suffix.
A precondition for correct functioning is that the file is of the expected type.
}
