% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilspec.R
\name{read.soilspec.with.suffix}
\alias{read.soilspec.with.suffix}
\title{Spectroscopy file reader function with assumed suffix explicitly supplied.
This function should be used when the file does not have the expected suffix.
A precondition for correct functioning is that the file and assumed suffix are compatible.}
\usage{
read.soilspec.with.suffix(
  path,
  assumed.suffix,
  is.absorbance = F,
  is.reflectance = F,
  is.transmittance = F,
  source.col.names = c("wavenumber", "intensity")
)
}
\arguments{
\item{path}{Full path to the file}

\item{assumed.suffix}{The assumed suffix for the expected file format, e.g. .spa, .0}

\item{is.absorbance}{Does the sample data correspond to absorbance?}

\item{is.reflectance}{Does the sample data correspond to reflectance?}

\item{is.transmittance}{Does the sample data correspond to transmittance?}
}
\value{
A result list containing a file read status, a data.frame of
wavenumber-intensity pairs, a list of any available metadata,
instrument mode, units, whether wavenumbers are in descending order,
instrument origin, spectrum type;
status will be non-zero if file does not exist or cannot be read (1),
is of an unknown format (2) or some other error occurred (4).
}
\description{
Spectroscopy file reader function with assumed suffix explicitly supplied.
This function should be used when the file does not have the expected suffix.
A precondition for correct functioning is that the file and assumed suffix are compatible.
}
