% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{spectral.evolution.file.path}
\alias{spectral.evolution.file.path}
\title{Return the full path to the Spectral Evolution example file}
\usage{
spectral.evolution.file.path()
}
\value{
The full path to the Spectral Evolution example file
}
\description{
Return the full path to the Spectral Evolution example file
}
