% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{thermo.spc.file.path}
\alias{thermo.spc.file.path}
\title{Return the full path to the Thermo spc example file}
\usage{
thermo.spc.file.path()
}
\value{
The full path to the Thermo spc example file
}
\description{
Return the full path to the Thermo spc example file
}
