% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{unknown.file.path}
\alias{unknown.file.path}
\title{Return the full path to the Unknown example file (for testing purposes)}
\usage{
unknown.file.path()
}
\value{
The full path to the Unknown example file
}
\description{
Return the full path to the Unknown example file (for testing purposes)
}
