% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraSelectionFunctions.R
\name{getSpectraSelectionInfo}
\alias{getSpectraSelectionInfo}
\title{Return information about the available spectral selection functions.}
\usage{
getSpectraSelectionInfo()
}
\arguments{
\item{type}{Transformation types to return  c('Preprocessing', 'Harmonisation')}
}
\value{
A dataframe of spectral selection information
}
\description{
Return information about the available spectral selection functions.
}
\examples{
getSpectraSelectionInfo()
}
